Level Manager working
Made exit-doors tile ID 15.

Can add levels. Next level will be loaded / can be switched to after a
level is beaten.

We don't have the intermediate screen that does this yet, though, so I
added a debug Spacebar that can be used to switch back to the gamescreen
from the timeout screen for now (pretending the timeout screen is the
loading/intermediate screen)